% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/podlove_baseplot_multi.R
\name{podlove_baseplot_multi}
\alias{podlove_baseplot_multi}
\title{Construct a basic podcast download ridgeline plot}
\usage{
podlove_baseplot_multi(df_tidy_data, gvar = NULL, cumulative = TRUE, ...)
}
\arguments{
\item{df_tidy_data}{A tidy data frame with download data, as constructed
by \code{podlove_prepare_stats_for_graph()}}

\item{gvar}{Optional grouping parameter string (e.g. "title"), handed over
to \code{ggplot2::aes_string(color)}. Must correspond to the data structure
in \code{df_tidy_data}.}

\item{cumulative}{Boolean switch to show either cumulative data (TRUE, default),
or non-comulative data (FALSE)}

\item{...}{Additional parameters to be handed over to \code{ggplot2::geom_density_ridge()}}
}
\value{
A ggplot2 plot
}
\description{
Creates the basic ridge line plot to display Podlove download data over time, using
    \code{ggplot2()}. Includes  an optional grouping variable and a switcher between
    cumulative/non-cumulative data.
}
\examples{
\dontrun{
#prepare data
gdata <- podlove_prepare_stats_for_graph(podcast_example_data, gvar = title)

# line graph with cumulative data (grouped by title)
podlove_baseplot(plot_data, gvar = "title", cumulative = TRUE)

# line graph with non-cumulative data
podlove_baseplot(plot_data, gvar = "title", cumulative = FALSE)

# line graph with non-cumulative data and additional formating paramters
gdata <- podlove_prepare_stats_for_graph(podcast_example_data)
podlove_baseplot(plot_data, cumulative = TRUE,
                 color = "white", fill = "#6699CC", scale = 2)
}

}
