% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/podlove_clean_stats.R
\name{podlove_clean_stats}
\alias{podlove_clean_stats}
\title{Connect and clean podcast and reference data}
\usage{
podlove_clean_stats(
  df_stats,
  df_mediafile,
  df_user,
  df_episodes,
  df_posts,
  launch_date = NULL
)
}
\arguments{
\item{df_stats}{contents of the MySQL table \code{wp_podlove_downloadintentclean}}

\item{df_mediafile}{contents of the MySQL table \code{wp_podlove_downloadintentclean}}

\item{df_user}{contents of the MySQL table \code{wp_podlove_useragent}}

\item{df_episodes}{contents of the MySQL table \code{wp_podlove_episode}}

\item{df_posts}{contents of the MySQL table \code{wp_posts}}

\item{launch_date}{date of the first official podcast episode release,
defaults to date of first download attempt (which may be before the first
episode was released)}
}
\value{
a dataframe containing all episode download attempts
}
\description{
\code{podlove_clean_stats} takes podcast download and various reference
data to create a tidy set of download attempts. The reference data consists 
of the tables for podlove episodes (\code{df_episodes}), blog posts 
(\code{df_posts}), user agents (\code{df_user}) and media files
(\code{df_mediafile}).
}
\examples{
\dontrun{
podlove_clean <- podlove_clean_stats(df_stats = table_stats,
                                     df_mediafile = table_mediafile,
                                     df_user = table_useragent,
                                     df_episodes = table_episodes,
                                     df_posts = table_posts,
                                     launch_date = "2017-12-04")
                                     }
}
