% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/podlove_create_example.R
\name{podlove_create_example}
\alias{podlove_create_example}
\title{Create Podcast Example Data}
\usage{
podlove_create_example(
  n_posts = 15,
  n_episodes = 10,
  n_useragents = 100,
  start_date = "2019-01-01",
  end_date = "2019-12-31",
  runtime = 30,
  total_dls = 5000,
  clean = FALSE,
  seed = NULL
)
}
\arguments{
\item{n_posts}{Number of posts to generate (defaults to 15)}

\item{n_episodes}{Number of episodes to generate (defaults to 12, must be 
lower or equal than n_posts)}

\item{n_useragents}{Number of useragents to sample from (defaults to 100)}

\item{start_date}{Podcast start date}

\item{end_date}{Podcast end date}

\item{runtime}{minimum runtime of an episode in days. This parameter defines 
the time difference between the release of the last episode and the last
download to be generated.}

\item{total_dls}{Total downloads to be generate. Defaults to 5000. Approximate value.}

\item{clean}{Switcher: if `TRUE`, cleans the data,
if `FALSE`, returns a list of tables}

\item{seed}{parameter to fix randomization via \code{set.seed()}}
}
\value{
A list of 5 named tables or a cleaned dataframe of downloads
}
\description{
This is a wrapper function to easily create example podcast download tables
    or a clean set of download data.
}
\examples{
# create tables for ~10000 downloads
test_list <- podlove_create_example(total_dls = 10000)

# create clean podcast data
test_df <- podlove_create_example(total_dls = 10000, clean = TRUE)

}
