% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/podlove_create_example_downloads.R
\name{podlove_create_example_downloads}
\alias{podlove_create_example_downloads}
\title{Create Random Podcast Download Data for a Set of Episodes}
\usage{
podlove_create_example_downloads(
  df_posts,
  df_episodes,
  df_mediafiles,
  df_useragents,
  runtime,
  dl_per_ep_day,
  seed
)
}
\arguments{
\item{df_posts}{data frame with example posts
(created by \code{podlove_create_example_posts()})}

\item{df_episodes}{data frame with example episodes
(created by \code{podlove_create_example_episodes()})}

\item{df_mediafiles}{data frame with example mediafiles
(created by \code{podlove_create_example_mediafiles()})}

\item{df_useragents}{edata frame with example mediafiles
(created by \code{podlove_create_example_useragents()})}

\item{runtime}{minimum runtime of an episode in days. This parameter defines 
the time difference between the release of the last episode and the last
download to be generated.}

\item{dl_per_ep_day}{Downloads per episode and day throughout its runtime. 
Be careful, this is a powerful parameter which will strongly increase the 
size of your dataset. Use the following formula to estimate this parameter:
dl_per_ep_day =  (2 \* total downloads) / (episodes \* total podcast runtime)
e.g. to create a set of 5000 downloads for 10 episodes over one year, use
(2 \* 5000 downloads) / (10 eps \* 365 days) = 2.74 downloads per ep/day.}

\item{seed}{parameter to fix randomization via \code{set.seed()}}
}
\value{
a tibble with downloads per episode
}
\description{
Lorem
}
\examples{
\dontrun{
test_posts <- podlove_create_example_posts(50, "2019-01-01", "2019-12-31")
test_episodes <- podlove_create_example_episodes(40, test_posts)
test_mediafiles <- podlove_create_example_mediafiles(test_episodes)
test_useragents <- podlove_create_example_useragents(100)
 
test_downloads <- podlove_create_example_downloads(test_posts,
                                                  test_episodes,
                                                  test_mediafiles,
                                                  test_useragents,
                                                  runtime = 100, 
                                                  dl_per_ep_day = 0.93)
}

}
