% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/podlove_create_example_episodes.R
\name{podlove_create_example_episodes}
\alias{podlove_create_example_episodes}
\title{Create a sample Wordpress "podlove_episodes" table}
\usage{
podlove_create_example_episodes(n_episodes, df_posts, seed = NULL)
}
\arguments{
\item{n_episodes}{number of posts (rows) to generate}

\item{df_posts}{data frame with corresponding posts}

\item{seed}{parameter to fix randomization via \code{set.seed()}}
}
\value{
a tibble with n_episodes episode entries and corresponding columns
}
\description{
This function generates a tibble with sample data corresponding to the Wordpress
    MySQL database table \code{wp_podlove_episodes}.
}
\examples{
# preparation: create posts 
example_posts <- podlove_create_example_posts(n_posts = 50,
                             start_date = "2018-10-25",
                             end_date = "2019-12-24")

# random posts episodes based
podlove_create_example_episodes(n_episodes = 30, 
                                df_posts = example_posts)

}
