% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/podlove_create_example_mediafiles.R
\name{podlove_create_example_mediafiles}
\alias{podlove_create_example_mediafiles}
\title{Create a sample Wordpress "podlove_mediafile" table}
\usage{
podlove_create_example_mediafiles(df_episodes, seed = NULL)
}
\arguments{
\item{df_episodes}{data frame with corresponding episodes}

\item{seed}{parameter to fix randomization via \code{set.seed()}}
}
\value{
a tibble with mediafile entries and corresponding columns
}
\description{
This function generates a tibble with sample data corresponding to the Wordpress
    MySQL database table \code{wp_podlove_mediafile}. Since this table is only
    used to connect episodes to download intents, the number of mediafiles is the
    same as the number of episodes (in reality, there are more mediafiles than
    episodes, as changes to the mediafiles can be made after release).
}
\examples{
# preparation: create posts & episodes 
example_posts <- podlove_create_example_posts(n_posts = 50,
                             start_date = "2018-10-25",
                             end_date = "2019-12-24")

example_eps <- podlove_create_example_episodes(n_episodes = 30, 
                                               df_posts = example_posts)
                                
# create mediafiles
podlove_create_example_mediafiles(df_episodes = example_eps)

}
