% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/podlove_create_example_posts.R
\name{podlove_create_example_posts}
\alias{podlove_create_example_posts}
\title{Create a sample Wordpress "posts" table}
\usage{
podlove_create_example_posts(n_posts, start_date, end_date, seed = NULL)
}
\arguments{
\item{n_posts}{number of posts (rows) to generate}

\item{start_date}{first post's publish date}

\item{end_date}{last post's publish date}

\item{seed}{parameter to fix randomization via \code{set.seed()}}
}
\value{
a tibble with n_posts post entries and corresponding columns
}
\description{
This function generates a tibble with sample data corresponding to the Wordpress
    MySQL database table \code{wp_posts}.
}
\examples{
# random posts
podlove_create_example_posts(n_posts = 20,
                             start_date = "2018-10-25",
                             end_date = "2019-12-24")

# random posts with fixed seed (try several times to see the results are equal)
podlove_create_example_posts(n_posts = 20, 
                             start_date = "2018-10-25", 
                             end_date = "2019-12-24",
                             seed = 10)

}
