% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/podlove_create_example_useragents.R
\name{podlove_create_example_useragents}
\alias{podlove_create_example_useragents}
\title{Create a sample Wordpress "podlove_useragent" table}
\usage{
podlove_create_example_useragents(
  n_useragents,
  df_ua_list = podcast_example_useragents,
  seed = NULL
)
}
\arguments{
\item{n_useragents}{number of posts (rows) to generate. can't be higher than
the number of entries in \code{df_ua_list}.}

\item{df_ua_list}{data frame with possible user agent combinations. defaults to
the in-built data set \code{podcast_example_useragents}}

\item{seed}{parameter to fix randomization via \code{set.seed()}}
}
\value{
a tibble with n_useragents rows
}
\description{
This function generates a tibble with sample data corresponding to the Wordpress
    MySQL database table \code{wp_podlove_useragent}.
}
\examples{
\dontrun{
# random useragent list 
example_useragents <- podlove_create_example_useragents(n_useragents = 100)
}

}
