% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/podlove_downloads_until.R
\name{podlove_downloads_until}
\alias{podlove_downloads_until}
\title{Calculate download sums at point(s) in time}
\usage{
podlove_downloads_until(df_tidy_data, points_in_time, time_unit = "days")
}
\arguments{
\item{df_tidy_data}{a tidy table of downloads created by \code{podlov_clean_stats()}
or \code{podlove_get_and_clean()}}

\item{points_in_time}{Measurement points in time relative to episode release
in hours or days. Can be a single value or a vector of integers.}

\item{time_unit}{Time unit of \code{points_in_time}, either \code{"days"} (default)
or \code{"hours"}.}
}
\value{
A tidy table of episode downloads at the specified points in time, 
    along with episode-related fields
}
\description{
This function takes a tidy table of downloads and creates a table containing
the number of downloads at specified \code{points_in_time} in hours or days.
Episodes with an episode age younger than \code{point_in_time} will be excluded.
Can be used to generate data for regression models.
}
\examples{
# downloads after 3 days
podlove_downloads_until(podcast_example_data, points_in_time = 3)

# downloads after 12 hours
podlove_downloads_until(podcast_example_data, points_in_time = 12,
                      time_unit = "hours")

# downloads after 1, 3, and 10 days
podlove_downloads_until(podcast_example_data, points_in_time = c(1, 3, 10))

}
