% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/podlove_episode_regression.R
\name{podlove_episode_regression}
\alias{podlove_episode_regression}
\title{TITLE}
\usage{
podlove_episode_regression(df_regression_data, terms = "post_datehour")
}
\arguments{
\item{lorem}{Ipsum}
}
\value{
Dolor
}
\description{
Description
}
\details{
Details
}
\examples{
1

}
