% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/podlove_graph_performance.R
\name{podlove_graph_performance}
\alias{podlove_graph_performance}
\title{Create performance grid for podcast episodes}
\usage{
podlove_graph_performance(
  df_perfstats,
  label = ep_number,
  legend = "",
  printout = TRUE
)
}
\arguments{
\item{df_perfstats}{a tidy data frame created by \code{performance_stats()}}

\item{label}{Unquoted, episode-related variable to use for labelling. By default,
\code{podlove_performance_stats} creates the options \code{title}, 
\code{ep_number} (default) and \code{ep_num_title}. Use \code{label = ""} to
display no label}

\item{legend}{Unquoted, episode-related variable to use in a explanatory legend 
next to the performance graph.}

\item{printout}{Switcher to automatically print out the plot (default TRUE)}
}
\value{
A ggplot object
}
\description{
Based on data created by \code{podlove_performance_stats}, plot all episodes
    on an X/Y grid, X showing long-term average downloads, Y showing average
    downloads during launch. Horizontal and vertcal lines show median values. 
    This allows for categorization of epsisodes into performance clusters.
    Note that you won't see episodes which are younger than your \code{post_launch}
    limit.
}
\examples{
\dontrun{
# plot episode performance stats with a launch period of 2 days and a
# post-launch period of 5 days

perf <- podlove_performance_stats(podcast_example_data, launch = 2, post_launch = 5)
podlove_graph_performance(perf)

# add a label
perf <- podlove_performance_stats(podcast_example_data, launch = 2, post_launch = 5,
                                  label = ep_number, legend = title)
podlove_graph_performance(perf)
}

}
