% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/podlove_graph_regression.R
\name{podlove_graph_regression}
\alias{podlove_graph_regression}
\title{Create a regression plot of episode downloads against an episode variable}
\usage{
podlove_graph_regression(
  df_regression_data,
  predictor = post_datehour,
  plot_type = "line",
  label = FALSE,
  regline = TRUE,
  ribbon = TRUE,
  stylize = TRUE,
  printout = TRUE
)
}
\arguments{
\item{df_regression_data}{a tidy data table created by \code{podlove_downloads_until()}.
Note that this function can't be fed data including more than one \code{point_in_time}}

\item{predictor}{a single predictor variable for downloads, typically a time 
or order based variable such as \code{post_date}, \code{post_datehour} (default), 
\code{episode_age_hours}/\code{episode_age_days} (note: age decreases
with episodes!) or \code{episode_rank}.}

\item{plot_type}{the type of plot that should be generated. accepts a string
from the list "point" (dot plot), "line" (line plot") or "area (area plot)}

\item{label}{boolean switcher to label the plotted points with episode titles
(default: FALSE)}

\item{regline}{boolean switcher if the plot should contain a regression line
(default: TRUE)}

\item{ribbon}{boolean switcher if the plot should contain a confidence ribbon
(default: TRUE)}

\item{stylize}{boolean switcher if the plot should be beautifed (default TRUE).
if FALSE, styling can be added later with \code{+ ggthemes::theme_name}}

\item{printout}{boolean switcher if the plot should print a graph}
}
\value{
a ggplot2 object
}
\description{
This function takes data from a \code{podlove_downloads_until()} and creates
    a regression plot with several graphical options.
}
\examples{
\dontrun{
# prepare linear regression data for downloads on day 3 by episode release date
dl <- podlove_downloads_until(podcast_example_data, points_in_time = 3)

# create dot plot 
podlove_graph_regression(dl, predictor = post_datehour, plot_type = "point")

# create line plot with labels 
podlove_graph_regression(dl, predictor = post_datehour, plot_type = "point",
                         label = TRUE)

# create area plot without regression line and ribbon 
podlove_graph_regression(dl, predictor = post_datehour, plot_type = "point",
                         regline = FALSE, ribbon = FALSE)

# create dot plot, add another theme 
podlove_graph_regression(dl, predictor = post_datehour, plot_type = "point",
                         stylize = FALSE) +
    ggthemes::theme_economist()
}

}
