% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/podlove_performance_stats.R
\name{podlove_performance_stats}
\alias{podlove_performance_stats}
\title{Calculate podcast episode performance indicators}
\usage{
podlove_performance_stats(
  df_tidy_data,
  launch = 3,
  post_launch = 7,
  limit_unit = "days"
)
}
\arguments{
\item{df_tidy_data}{A tidy data frame with download data, as constructed
by \code{podlove_clean_stats()} or \code{podlove_get_and_clean}.}

\item{launch}{definition of a episode launch period in days after launch}

\item{post_launch}{definition of begin of long-term performance in days after launch}

\item{limit_unit}{time unit for limits. Can be "days" (default) or "hours".
Used to fine-tune launch performance cutoffs.}
}
\value{
a dataframe containing performance data per episode title (total
    total downloads, average downloads during launch, average downloads 
    after launch)
}
\description{
Calculates totals and average downloads of podcast episodes with flexible
    defitions of "launch" and "evergreen" (long-term) periods.
    
Podcast performance between episodes is difficult to measure and compare,
    because podcasts a) have in most cases successive release dates, 
    b) follow non-linear, usually exponential performance patterns. Comparisons
    therefore need to be based on indicators reflecting those peculiarities. 
    
    One possible approach is to look at average downloads per time unit (e.g.
    days), and split performance in an initial "launch" period and a long-
    term "evergreen" period. The function \code{podlove_performance_stats} 
    and its helper function \code{podlove_total_average_downloads} allow this.
}
\examples{
# performance stats with a launch period of 2 days and a post-launch period
# of 5 days
podlove_performance_stats(podcast_example_data, launch = 2, post_launch = 5)

# performance stats with a launch period of 12 hours and a post-launch period
# of 120 hours (= 5 days)
podlove_performance_stats(podcast_example_data,
                          launch = 2, post_launch = 5*24, limit_unit = "hours")

}
\seealso{
\code{podlove_total_average_downloads} for a more the helper function
   behind this function, which allows more fine tuned data.
}
