% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/podlove_graph_download_curves.R
\name{podlove_plot_download_curves}
\alias{podlove_plot_download_curves}
\title{Create download curve plot}
\usage{
podlove_plot_download_curves(
  df_tidy_data,
  gvar = NULL,
  cumulative = TRUE,
  labelmethod = "last.points",
  printout = TRUE
)
}
\arguments{
\item{df_tidy_data}{A tidy data frame with download data, as constructed
by \code{podlove_prepare_stats_for_graph()}}

\item{gvar}{Optional grouping parameter (e.g. episode title), handed over
to \code{ggplot2::aes(color)}. Must be quoted (e.g. \code{"title"})}

\item{cumulative}{Boolean switch to show either cumulative data (TRUE, default),
or non-comulative data (FALSE)}

\item{labelmethod}{Defines where should the labels be attached 
(at the beinning of the curves: \code{"last.points"}, default, or at
the end of the curves \code{"first.points"})}

\item{printout}{Switcher to automatically print out the plot (default TRUE)}
}
\value{
A ggplot object
}
\description{
Based on a prepared dataset, this function creates a line curve diagram
    of downloads over time. It takes an optional grouping variable (e.g.
    title) as well as options for cumulative display, labeling and 
    print output. For more finetuning options, use the \code{podlove_baseplot()}
    function and add geoms to your liking.
}
\examples{
\dontrun{
}
}
