% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/podlove_plot_regression.R
\name{podlove_plot_regression}
\alias{podlove_plot_regression}
\title{Create a regression plot of episode downloads against an episode variable}
\usage{
podlove_plot_regression(
  df_tidy_data,
  point_in_time,
  time_unit = "days",
  predictor = post_datehour,
  plot_type = "line",
  label = FALSE,
  regline = TRUE,
  ribbon = TRUE,
  stylize = TRUE,
  print_model = TRUE,
  print_plot = TRUE
)
}
\arguments{
\item{df_tidy_data}{a tidy table of downloads created by \code{podlov_clean_stats()}
or \code{podlove_get_and_clean()}}

\item{point_in_time}{Measurement point in time relative to episode release
in hours or days. Can be a single value or a vector of integers.}

\item{time_unit}{Time unit of \code{points_in_time}, either \code{"days"} (default)
or \code{"hours"}.}

\item{predictor}{a single predictor variable for downloads, typically a time 
or order based variable such as \code{post_date}, \code{post_datehour} (default), 
\code{episode_age_hours}/\code{episode_age_days} (note: age decreases
with episodes!) or \code{episode_rank}.}

\item{plot_type}{the type of plot that should be generated. accepts a string
from the list "point" (dot plot), "line" (line plot") or "area (area plot)}

\item{label}{boolean switcher to label the plotted points with episode titles
(default: FALSE)}

\item{regline}{boolean switcher if the plot should contain a regression line
(default: TRUE)}

\item{ribbon}{boolean switcher if the plot should contain a confidence ribbon
(default: TRUE)}

\item{stylize}{boolean switcher if the plot should be beautifed (default TRUE).
if FALSE, styling can be added later with \code{+ ggthemes::theme_name}}

\item{print_model}{boolean switcher if the regression model should be printed}

\item{print_plot}{boolean switcher if the plot should print a graph}
}
\value{
a ggplot2 object
}
\description{
Wrapper function to take data from \code{podlove_get_and_clean()} and creates
    a regression plot with several graphical options.
}
\examples{
\dontrun{
# prepare linear regression data for downloads on day 3 by episode release date
downloads <- podlove_create_example()

# create dot plot 
podlove_plot_regression(downloads, point_in_time = 3, plot_type = "point")

# create line plot with labels 
podlove_plot_regression(downloads, point_in_time = 3, plot_type = "point",
                         label = TRUE)

# create area plot without regression line and ribbon 
podlove_plot_regression(downloads, point_in_time = 3, predictor = post_datehour, 
                         plot_type = "point", regline = FALSE, ribbon = FALSE)

# create dot plot, add another theme 
podlove_plot_regression(downloads, point_in_time = 3, predictor = post_datehour, 
                        plot_type = "point", stylize = FALSE) +
    ggthemes::theme_economist()
}

}
