% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/podlove_podcast_summary.R
\name{podlove_podcast_summary}
\alias{podlove_podcast_summary}
\title{Summarize Podcast data}
\usage{
podlove_podcast_summary(df_tidy_data, verbose = TRUE, return_params = FALSE)
}
\description{
Calculates, returns and optionally downloads per episode
    `downloads_per_episode_5num`: five-number summary of downloads per episode
    `downloads_per_day_mean`: average downloads per day
    `downloads_per_day_5num`: five-number summary of downloads per day
}
\examples{
#create sample data
dls <- podlove_create_example(clean = TRUE)

#get summary
dls_summary <- podlove_podcast_summary(dls) 

}
