% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/podlove_prep_regtable.R
\name{podlove_prep_regtable}
\alias{podlove_prep_regtable}
\title{Calculate download sums at point in time}
\usage{
podlove_prep_regtable(df_tidy_data, point_in_time)
}
\arguments{
\item{df_tidy_data}{a tidy table of downloads created by \code{podlov_clean_stats()}
or \code{podlove_get_and_clean()}}

\item{point_in_time}{Measurement point in time relative to episode release
in hours. To specify days, use \code{point_in_time = days * 24}}
}
\value{
A tidy table of episode downloads at the specified point in time, 
    along with episode-related fields
}
\description{
This function takes a tidy table of downloads and creates a table containing
the number of downloads at a specified \code{point_in_time} in hours.
Episodes with an episode age younger than \code{point_in_time} will be excluded.
Helper function to \code{podlove_prep_regtables()}.
}
\examples{
\dontrun{
# downloads after 12 hours
podlove_prep_regtable(podcast_example_data, 12)

# downloads after 10 days
podlove_prep_regtable(podcast_example_data, 10*24)
}

}
