% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/podlove_prepare_stats_for_graph.R
\name{podlove_prepare_stats_for_graph}
\alias{podlove_prepare_stats_for_graph}
\title{Prepare Podlove data for graphical rendering}
\usage{
podlove_prepare_stats_for_graph(
  df_stats,
  gvar,
  hourly = FALSE,
  relative = TRUE,
  last_n = 0
)
}
\arguments{
\item{df_stats}{A data frame of clean Podlove data, as rendered by 
\code{podlove_get_and_clean()} or \code{podlove_clean_stats}.}

\item{gvar}{Optional grouping variable(s), unquoted.}

\item{hourly}{Boolean switching parameter for rendering of hourly vs.
daily data. Defaults to \code{FALSE} (daily data), \code{TRUE} creates
hourly data.}

\item{relative}{Boolean switching parameter to define if the data is 
rendered relative to the respective episode release date (\code{TRUE}) or
in absolute dates (\code{TRUE}). Defaults to \code{TRUE}.}

\item{last_n}{Number of most recent episodes to filter for. Defaults to 0 (no filtering),
use negative numbers to filter for first n episodes.}
}
\description{
Takes clean Podlove data and prepares it to be rendered with \code{ggplot2}. 
The function accepts one or more grouping variables (e.g. episode title) and
accepts parameters for the time axis of downloads.
}
\examples{
\dontrun{
# relative, daily plot by episode title
podlove_prepare_stats_for_graph(podcast_example_data, gvar = title) 

# relative, hourly plot by episode number
podlove_prepare_stats_for_graph(podcast_example_data, gvar = ep_number, hourly = TRUE) 

# absolute, daily plot by episode title
podlove_prepare_stats_for_graph(podcast_example_data, gvar = title, relative = FALSE) 

# abolute, hourly plot by podcast client name
podlove_prepare_stats_for_graph(podcast_example_data, gvar = client_name, relative = FALSE) 
}

}
