% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/podlove_total_average_downloads.R
\name{podlove_total_average_downloads}
\alias{podlove_total_average_downloads}
\title{Calculate total downloads and average downloads over time}
\usage{
podlove_total_average_downloads(
  df_tidy_data,
  gvar = "title",
  lower_limit = 0,
  upper_limit = Inf,
  limit_unit = "days"
)
}
\arguments{
\item{df_tidy_data}{A tidy data frame with download data, as constructed
by \code{podlove_clean_stats()} or \code{podlove_get_and_clean}.}

\item{gvar}{an optional grouping variable, which should be reflect the episode
structure (e.g. ep_number, title). Other variables will lead to nonsensical
results, because the calculation is based on time unit since episode release.}

\item{lower_limit}{lower cutoff for time since episode launch (used to eliminate
launch peaks). Defaults to 0, i.e. no filtering.}

\item{upper_limit}{upper cutoff for time since episode launch (used to eliminate
long term download data and just focus on launch performance)}

\item{limit_unit}{time unit for limits. Can be "days" (default) or "hours".
Used to fine-tune launch performance cutoffs.}
}
\value{
a dataframe containing performance data per episode (total downloads,
    during time period, average downloads during time period).
}
\description{
Calculates the total of downloads and the average of downloads over a time 
    unit for a podcast or its episodes. The measurement period can be defined
    with an upper and lower limit. Use this function for fine-tuned performance
    measurement.
}
\examples{
\dontrun{
# total and average downloads by episode, no timeframe filter (complete period)
podlove_total_average_downloads(podcast_example_data, gvar = "title")

# total and average downloads by episode number after 7 days of release 
podlove_total_average_downloads(podcast_example_data, gvar = "ep_number",
                               lower_limit = 7)

# total and average downloads by episode number during first 12 hours of release 
podlove_total_average_downloads(podcast_example_data, gvar = "ep_number",
                               upper_limit = 12, limit_unit = "hours")
                               }

}
\seealso{
\code{podlove_performance_stats} for a more general performance function
}
