% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/podlove_baseplot.R
\name{podlove_baseplot}
\alias{podlove_baseplot}
\title{Construct a basic Podlove plot 2}
\usage{
podlove_baseplot(df_tidy_data, gvar = NULL, cumulative = TRUE, ...)
}
\arguments{
\item{df_tidy_data}{A tidy data frame with download data, as constructed
by \code{podlove_prepare_stats_for_graph()}}

\item{gvar}{Optional grouping parameter (e.g. episode title), handed over
to \code{ggplot2::aes(color)}.}

\item{cumulative}{Boolean switch to show either cumulative data (TRUE, default),
or non-comulative data (FALSE)}

\item{...}{Additional parameters to be handed over to \code{ggplot2::aes()}}
}
\value{
A ggplot2 plot
}
\description{
Creates the basic plot to display Podlove download data over time, using
    \code{ggplot2()}. Does not add any geoms except axis limits. Includes 
    an optional grouping variable and a switcher between
    cumulative/non-cumulative data.
}
\examples{
\dontrun{
print(1)
}

}
