% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/podlove_get_and_clean.R
\name{podlove_get_and_clean}
\alias{podlove_get_and_clean}
\title{Fetch a table from a MySQL database}
\usage{
podlove_get_and_clean(
  db_name = rstudioapi::askForSecret(name = "dbname"),
  db_host = rstudioapi::askForSecret(name = "host"),
  db_user = rstudioapi::askForSecret(name = "user"),
  db_password = rstudioapi::askForSecret(name = "password"),
  launch_date
)
}
\arguments{
\item{tablename}{name of the MySQL table}

\item{dbname}{name of the database}

\item{host}{hostname of the database}

\item{user}{username of the database}

\item{password}{password of the database}
}
\value{
A dataframe containing the table data.
}
\description{
\code{podlove_get_table_data} opens up a connection to a MySQL database and
fetches a whole table as a df. It Uses \code{rstudioapi::askForSecret()}, 
which allows securely storing access info in the local keyring.
}
\examples{
\dontrun{
dlic_raw <- podlove_get_table_data(db_stats)
}
}
