% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/podlove_plot_curves.R
\name{podlove_plot_curves}
\alias{podlove_plot_curves}
\title{Plot Download Curves from Download Data}
\usage{
podlove_plot_curves(
  dldata,
  gvar = "Total",
  hourly = FALSE,
  relative = TRUE,
  cumulative = TRUE,
  plot_type = "line",
  labelmethod = "last.points",
  printout = TRUE,
  last_n = 0,
  ...
)
}
\arguments{
\item{dldata}{A tidy data frame with download data, as constructed
by \code{podlove_prepare_stats_for_graph()}}

\item{gvar}{Optional grouping parameter (e.g. episode title), handed over
to \code{ggplot2::aes(color)}.}

\item{hourly}{Boolean switching parameter for rendering of hourly vs.
daily data. Defaults to \code{FALSE} (daily data), \code{TRUE} creates
hourly data.}

\item{relative}{Boolean switching parameter to define if the data is 
rendered relative to the respective episode release date (\code{TRUE}) or
in absolute dates (\code{TRUE}). Defaults to \code{TRUE}.}

\item{cumulative}{Boolean switch to show either cumulative data (TRUE, default),
or non-comulative data (FALSE)}

\item{plot_type}{Sets the plot type to either line plot (\code{"line"}) or
ridgeline plot (\code{"ridge"}).}

\item{labelmethod}{Defines where should the labels be attached 
(at the beinning of the curves: \code{"last.points"}, default, or at
the end of the curves \code{"first.points"})}

\item{printout}{Switcher to automatically print out the plot (default TRUE)}

\item{last_n}{Number of most recent episodes to filter for. Defaults to 0 
(no filtering), use negative numbers to filter for first n episodes.}

\item{...}{additional formating parameters for \code{ggplot2::geom_line()}
or \code{ggridges::geom_density_ridges}.}
}
\value{
A ggplot object
}
\description{
Wrapper function to prepare and generate a line curve diagram
    of downloads over time. It takes an optional grouping variable (e.g.
    title) as well as options for the time axis of downloads for cumulative display, 
    labeling and print output. If you want to generate and tweak the data
    behind this plot, use \code{podlove_prepare_stats_for_graph}.
}
\examples{
# create example data
dls <- podlove_create_example(clean = TRUE)

podlove_plot_curves(dls, gvar = title, relative = TRUE, cumulative = TRUE)

# see podlove_graph_download_curves() for more information.

}
\seealso{
podlove_graph_download_curves(), 
   podlove_prepare_stats_for_graph()
}
